wip(text): text path edit
