+++ src/editor/textbox.tsx
@@ -3,7 +3,29 @@ import { TEXTBOX_DEFAULT_CONFIG } from '../utils/constants';
 import { uuid, loadFont } from '@/utils';
 import { getGlobalEditor } from '@/utils/global';
 import { setObject2Center } from '@/utils/helper';
-import { drawTextPath } from './line';
+
+export const getTextboxWidth = (textbox) => {
+  const textLines = textbox.textLines || [];
+  if (!textLines || !textLines.length) return 0;
+  let width = 0;
+  for (let i = 0; i < textLines.length; i++) {
+    width += textbox.measureLine(i).width;
+  }
+  return width;
+}
+
+export const drawTextPath = (textbox, offset) => {
+  const width = textbox.width;
+  const pathStr = `M 0 0 Q ${width / 2} ${offset} ${width} 0`;
+  const path = new fabric.Path(pathStr, {
+    visible: false,
+    stroke: '#000000',
+    fill: '#00000000'
+  });
+
+  textbox.set('path', path);
+  textbox.set('_forceClearCache', true);
+}
 
 export const createTextbox = async (options) => {
   const { text = '', left, top, fontFamily, ...rest } = options || {};
@@ -36,7 +58,6 @@ export const createTextbox = async (options) => {
     if (textBox.path) {
       hasPath = true;
       textBox.set('path', undefined);
-      textBox.setCoords();
       editor.canvas.requestRenderAll();
     } else {
       hasPath = false;
@@ -45,9 +66,10 @@ export const createTextbox = async (options) => {
 
   textBox.on('editing:exited', () => {
     if (hasPath) {
-      const _path = drawTextPath(textBox, Math.floor(textBox.width / 2));
-      textBox.set('path', _path);
-      textBox.setCoords();
+      const width = getTextboxWidth(textBox) + 10;
+      textBox.set('width', width);
+      editor.canvas.requestRenderAll();
+      drawTextPath(textBox, Math.floor(width / 2));
       editor.canvas.requestRenderAll();
     }
   });
